01 - 01 - 25
Use a helper function to push left elements of the tree

                                  APPROACH
The problem is solved using a stack to simulate an in-order traversal of the BST. The stack stores nodes along the left path, and the next() method pops the smallest node, pushing its right child’s left subtree onto the stack for future visits. This ensures we return nodes in sorted order.